requirements.txt
pyyaml==6.0
colorama==0.4.4
keyboard==0.13.5